only allow admin to reject events

+++ b13/hoos/views.py
@@ -74,6 +74,8 @@ def approve_event(request, event_id):
 def reject_event(request, event_id):
     if not request.user.is_authenticated:
         return redirect('home')
+    if not is_admin(request.user):
+        return redirect('home')
     event = EventSubmission.objects.get(id=event_id)
     event.delete()
     return redirect('review_events')
